THRIFT-5978: Apply declare(strict_types=1) in PHP runtime library (#3457)
Client: php

Add declare(strict_types=1); to every PHP file in lib/php/lib/ and
non-fixture lib/php/test/, between the license docblock and the namespace.
Strict mode now enforces the native parameter and property types added in
THRIFT-5976 / THRIFT-5977 — silent scalar coercion ("5" -> 5, 0.5 -> 0)
becomes a TypeError at the call site instead of slipping through.

Coercion bugs surfaced and fixed:

* TPhpStream::__construct cast bitwise int-result of $mode & MODE_R/MODE_W
  into typed bool $read/$write properties; now explicit (bool) cast and
  int $mode parameter.
* TJSONProtocol::writeJSONInteger / TSimpleJSONProtocol::writeJSONInteger
  wrote the int straight into the transport (which calls strlen()); now
  cast to (string) before writing and the parameter is typed int.
* TJSONProtocol::writeJSONString manually wraps numeric values in QUOTE
  for escapeNum contexts but then ran them through json_encode, which
  added its own quotes for numeric-string inputs ("0" -> ""0"" instead
  of "0"). Replaced the wrap+json_encode path with a direct numeric
  emission so int and numeric-string keys round-trip identically.
* writeString is now typed string across the protocol hierarchy
  (TProtocol abstract + TBinaryProtocol, TCompactProtocol, TJSONProtocol,
  TSimpleJSONProtocol concrete + TProtocolDecorator). PHP map keys that
  are numeric strings collapse to int internally; loose-mode generated
  callers now coerce them at the call boundary instead of fataling on
  strlen() inside the strict callee.
* TSocketPool::addServer is typed (string $host, int $port); the
  constructor now delegates to it for the array-builder loop, so storage
  shape and validation live in one place. test/php/TestClient.php and
  test/php/TestServer.php parse --port=NNNN via substr() / getopt, which
  return string; cast to int at the parse site so loose callers reach
  the typed constructor with the right type.
* THttpClientTest::flushDataProvider declared 'port' => '80' (string);
  THttpClient's promoted $port is int -- replaced with literal 80.

Generated fixtures under lib/php/test/Resources/packages/ stay untyped
(generator regen tracked separately).

Generated-by: Claude Opus 4.7

+++ test/php/TestClient.php
@@ -90,7 +90,7 @@ if ($argc > 2) {
 
 foreach ($argv as $arg) {
   if (substr($arg, 0, 7) == '--port=') {
-    $port = substr($arg, 7);
+    $port = (int) substr($arg, 7);
   } else if (substr($arg, 0, 12) == '--transport=') {
     $MODE = substr($arg, 12);
   } else if (substr($arg, 0, 11) == '--protocol=') {
